added logic to throw error when deleting scaling or analyzer or oxide that are connected to any channel and also added logic to delete station and channel when deleted company to prevent showing up anywhere

+++ Services/Services/AnalyzerService.cs
@@ -3,15 +3,18 @@ using Post= Models.Post;
 using Repositories.Interfaces;
 using Services.Interfaces;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Services
 {
     public class AnalyzerService : IAnalyzerService
     {
         private readonly IAnalyzerRepository _analyzerRepository;
-        public AnalyzerService(IAnalyzerRepository analyzerRepository)
+        private readonly IChannelRepository _channelRepository;
+        public AnalyzerService(IAnalyzerRepository analyzerRepository, IChannelRepository channelRepository)
         {
             _analyzerRepository = analyzerRepository;
+            _channelRepository = channelRepository;
         }
         public void CreateAnalyzer(Post.Analyzer analyzer)
         {
@@ -20,6 +23,12 @@ namespace Services
 
         public void DeleteAnalyzer(int id)
         {
+            var channelsLinkedToAnalyzer = _channelRepository.GetAll().Where(e => e.ProtocolId == id).ToList();
+            if (channelsLinkedToAnalyzer.Any())
+            {
+                var analyzer = _analyzerRepository.GetById(id);
+                throw new Exceptions.AnalyzerCannotBeDeletedException(analyzer.ProtocolType, string.Join(",", channelsLinkedToAnalyzer.Select(e => e.Name)));
+            }
             _analyzerRepository.Delete(id);
         }
