+++ Services/Services/OxideService.cs
@@ -3,16 +3,19 @@ using Post = Models.Post;
 using Repositories.Interfaces;
 using Services.Interfaces;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Services
 {
     public class OxideService : IOxideService
     {
         private readonly IOxideRepository _oxideRepository;
+        private readonly IChannelRepository _channelRepository;
 
-        public OxideService(IOxideRepository oxideRepository)
+        public OxideService(IOxideRepository oxideRepository, IChannelRepository channelRepository)
         {
             _oxideRepository = oxideRepository;
+            _channelRepository = channelRepository;
         }
         public void CreateOxide(Post.Oxide oxide)
         {
@@ -21,6 +24,12 @@ namespace Services
 
         public void DeleteOxide(int id)
         {
+            var channelsLinkedToTheOxide = _channelRepository.GetAll().Where(e => e.OxideId == id).ToList();
+            if (channelsLinkedToTheOxide.Any())
+            {
+                var oxide = _oxideRepository.GetById(id);
+                throw new Exceptions.OxideCannotBeDeletedException(oxide.OxideName, string.Join(",", channelsLinkedToTheOxide.Select(e => e.Name)));
+            }
             _oxideRepository.Delete(id);
         }
